fix(admin): 채팅방 관리 페이지 리팩토링
